Fixed A Lot Of Bugs, Converter, transaction make and revert :)

+++ src/main/java/com/example/tourism_management_system/bank/api/service/impl/CardServiceImpl.java
@@ -92,7 +92,7 @@ public class CardServiceImpl implements CardService {
     public String withdrawBalance(String cardNumber, double balance) {
         Optional<CardEntity> cardEntity = cardRepository.findCardEntityByCardNumber(cardNumber);
         if (balance <= cardEntity.get().getBalance() && balance >= 1) {
-            cardEntity.get().setBalance((float) (cardEntity.get().getBalance() - balance));
+            cardEntity.get().setBalance((cardEntity.get().getBalance() - balance));
             cardRepository.save(cardEntity.get());
         } else return "You don't have enough money to complete the transaction";
         return "Transaction completed successfully";
